Handle larger than machine strides in convolution C++ vertices
Summary:
Use 21-bit strides so that packed strides fit in a 64-bit
vertex field. The stride unpacking function always takes
64-bit packed stride even for 10-bit strides. But this is
not seen as a problem because the function to unpack
strides should be called for C++ vertices where strides
are 21-bit. Otherwise an overload to unpack 32-bit strides
could be added.

QUARTER type is not handled properly primarily because
unlimited version is not properly supported for that
type.

Ref T62097

Reviewers: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: timp, #poplar_team, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T62097

Differential Revision: https://phabricator.sourcevertex.net/D66884

+++ lib/poplin/ConvPartialsStridesPacking.cpp
@@ -68,7 +68,8 @@ std::pair<bool, int> reverseTransfromedOutStride(int transformedOutStride,
   return std::make_pair(flipOut, outStride);
 }
 
-unsigned packAmpNx1Stride(const unsigned strideBits, int p2, int p1, int p0) {
+unsigned long long packAmpNx1Stride(const unsigned strideBits, int p2, int p1,
+                                    int p0) {
   const unsigned strideMask = (1 << strideBits) - 1;
   const unsigned s10 = strideBits;
   const unsigned s20 = 2 * strideBits;
@@ -77,8 +78,8 @@ unsigned packAmpNx1Stride(const unsigned strideBits, int p2, int p1, int p0) {
          (p0 & strideMask);
 }
 
-int unpackAmpNx1Stride(const unsigned strideBits, unsigned stride, unsigned pX,
-                       bool signExtention) {
+int unpackAmpNx1Stride(const unsigned strideBits, unsigned long long stride,
+                       unsigned pX, bool signExtention) {
   const unsigned strideMask = (1 << strideBits) - 1;
   const unsigned shift = pX * strideBits;
   const int mask = 1 << (strideBits - 1);
